requirements.txt
# 기존 의존성
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0
requests==2.31.0

# 새로운 인증 시스템 의존성
aiomysql==0.2.0
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8
email-validator==2.1.0

# 개발 및 테스팅 (선택적)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2